Resolve incorrect negative survival time values in the Go port (#18)
* Update configuration and test result display to reduce clutter

Brings many of the descriptions closer to the Python port to make comparing test cases easier.

* Bring formatting closer to the rust / python version, add test cases

* Add baseline test, fix last formatting issue

* Fix didn't die issue with negative actualDPS

Same fix as the Python version

* Optimized 1.0 calculations

There were tight loops that were effectively calculating 1.0 - foo every test case. This has been optimized to be done once per loadout or generator.

* Bug fixes to make the optimization actually return the same values as PS and Python versions

The rust version is still not quite right for negative DPS, but this output does run correctly now with the various test cases. Also, updated output to more closely match that of the Rust version which is a bit more compact and readable.